feat: allow users to move cart item back to wishlist

+++ src/app/slices/addressAPI.js
@@ -1,4 +1,6 @@
 import axios from 'axios';
+import store from '../store';
+import { selectAddresses, selectDefaultAddress, selectAddressById } from './addressSlice';
 
 const client = axios.create({
   baseURL: 'http://localhost:8000',
@@ -15,6 +17,28 @@ export async function fetchAddressesAPI(userId) {
 }
 
 export async function addNewAddressAPI(address, userId) {
+  const addresses = selectAddresses(store.getState());
+  const defaultAddress = selectDefaultAddress(store.getState());
+
+  if (addresses.length === 0) {
+    address.default = true;
+  }
+
+  if (address.default) {
+    const config = {
+      method: 'patch',
+      url: `/addresses/${defaultAddress.id}`,
+      data: {
+        default: false,
+      },
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    };
+
+    await client(config);
+  }
+
   const config = {
     method: 'post',
     url: '/addresses',
@@ -32,6 +56,23 @@ export async function addNewAddressAPI(address, userId) {
 }
 
 export async function updateAddressAPI(id, updates) {
+  if (updates.default) {
+    const defaultAddress = selectDefaultAddress(store.getState());
+
+    const config = {
+      method: 'patch',
+      url: `/addresses/${defaultAddress.id}`,
+      data: {
+        default: false,
+      },
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    };
+
+    await client(config);
+  }
+
   const config = {
     method: 'patch',
     url: `/addresses/${id}`,
@@ -48,6 +89,12 @@ export async function updateAddressAPI(id, updates) {
 }
 
 export async function deleteAddressAPI(id) {
+  const address = selectAddressById(store.getState(), id);
+
+  if (address.default) {
+    throw new Error("Default address can't be deleted");
+  }
+
   const config = {
     method: 'delete',
     url: `/addresses/${id}`,
@@ -55,4 +102,35 @@ export async function deleteAddressAPI(id) {
 
   await client(config);
   return id;
-};
+}
+
+export async function setAsDefaultAPI(id) {
+  const defaultAddress = selectDefaultAddress(store.getState());
+
+  let config = {
+    method: 'patch',
+    url: `/addresses/${defaultAddress.id}`,
+    data: {
+      default: false,
+    },
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  };
+
+  await client(config);
+
+  config = {
+    method: 'patch',
+    url: `/addresses/${id}`,
+    data: {
+      default: true,
+    },
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  };
+
+  const response = await client(config);
+  return response.data;
+}
